requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
cairocffi==1.4.0
CairoSVG==2.6.0
cffi==1.15.1
charset-normalizer==3.1.0
chess==1.9.4
colorama==0.4.6
comm==0.1.2
cssselect2==0.7.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
discord-py-interactions==4.4.0
discord.py==2.2.2
exceptiongroup==1.1.1
executing==1.2.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==305
pyzmq==25.0.1
six==1.16.0
stack-data==0.6.2
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
